# backend/docs/swagger.yaml
definitions:
  apm.QueryResult:
    properties:
      hits:
        items:
          additionalProperties: {}
          type: object
        type: array
      total:
        type: integer
    type: object
...
  domain.EventKey:
    enum:
    - login
    - logout
    - beforeUpdateArticle
    - afterUpdateArticle
    - deleteArticle
    - beforeUpdateDraft
    - afterUpdateDraft
    - deleteDraft
    - updateSiteInfo
    - manualTriggerEvent
    type: string
    x-enum-varnames:
    - LoginEvent
    - LogoutEvent
    - BeforeUpdateArticleEvent
    - AfterUpdateArticleEvent
    - DeleteArticleEvent
    - BeforeUpdateDraftEvent
    - AfterUpdateDraftEvent
    - DeleteDraftEvent
    - UpdateSiteInfoEvent
    - ManualTriggerEvent
  domain.EventType:
    enum:
    - system
    - custom
    - cron
    type: string
    x-enum-varnames:
    - SystemEventType
    - CustomEventType
    - CornEventType
...
  domain.MenuItem:
    properties:
      children:
        $ref: '#/definitions/domain.MenuItem'
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      value:
        type: string
    type: object
...
  domain.Pipeline:
    properties:
      createdAt:
        type: string
      deleted:
        type: boolean
      enabled:
        type: boolean
      eventName:
        $ref: '#/definitions/domain.EventKey'
      eventType:
        $ref: '#/definitions/domain.EventType'
      id:
        type: integer
      name:
        type: string
      script:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/log:
    get:
      consumes:
      - application/json
      description: get logs
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: eventType
        in: query
        name: eventType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apm.QueryResult'
      summary: get logs
      tags:
      - Log
...
  /api/admin/meta/menu:
    get:
      consumes:
      - application/json
      description: get menu by settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.MenuItem'
            type: array
      summary: get menu by settings
      tags:
      - Menu
    put:
      consumes:
      - application/json
      description: update menu
      parameters:
      - description: menu
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/domain.MenuItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update menu
      tags:
      - Menu
...
  /api/admin/pipeline:
    get:
      consumes:
      - application/json
      description: get all pipeline
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Pipeline'
            type: array
      summary: get all pipeline
      tags:
      - Pipeline
    post:
      consumes:
      - application/json
      description: create pipeline
      parameters:
      - description: pipeline
        in: body
        name: pipeline
        required: true
        schema:
          $ref: '#/definitions/domain.Pipeline'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: create pipeline
      tags:
      - Pipeline
    put:
      consumes:
      - application/json
      description: update pipeline
      parameters:
      - description: pipeline
        in: body
        name: pipeline
        required: true
        schema:
          $ref: '#/definitions/domain.Pipeline'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update pipeline
      tags:
      - Pipeline
...
swagger: "2.0"